ワールド編集とhandlerのリファクタリング
